cv_rag_project/requirement.txt
streamlit
pandas 
pypdf 
python-docx
langchain
langchain-community
langchain-ollama
langchain-text-splitters
langchain-chroma
chromadb 

